Moved some query stats back into handler. Fixed processlist tests. Removed a spurious debug from a test
